refactor: rename contacts to contact
