+++ src/app/i18n/helpers/index.ts
@@ -1,5 +1,5 @@
 // Translations
-import { translations as CONTACTS } from "../../../domains/contacts/i18n";
+import { translations as CONTACTS } from "../../../domains/contact/i18n";
 import { translations as EXCHANGE } from "../../../domains/exchange/i18n";
 import { translations as HELP } from "../../../domains/help/i18n";
 import { translations as OFFLINE } from "../../../domains/offline/i18n";
