+++ src/domains/contact/components/DeleteContact/index.ts
