+++ src/domains/contact/pages/Contacts/Contacts.tsx
@@ -5,15 +5,15 @@ import { Button } from "app/components/Button";
 import { Header } from "app/components/Header";
 import { HeaderSearchBar } from "app/components/Header/HeaderSearchBar";
 import { Table } from "app/components/Table";
-import { ContactListItem } from "domains/contacts/components/ContactListItem";
-import { CreateContact } from "domains/contacts/components/CreateContact";
+import { ContactListItem } from "domains/contact/components/ContactListItem";
+import { CreateContact } from "domains/contact/components/CreateContact";
 import React, { useState } from "react";
 import { useTranslation } from "react-i18next";
 
 const { ContactsBanner } = images.contacts.pages.contacts;
 
 type ContactsHeaderExtraProps = {
-  showSearchBar: boolean;
+	showSearchBar: boolean;
 	onSearch?: any;
 	onAddContact?: any;
 };
@@ -21,17 +21,18 @@ type ContactsHeaderExtraProps = {
 const ContactsHeaderExtra = ({ showSearchBar, onSearch, onAddContact }: ContactsHeaderExtraProps) => {
 	const { t } = useTranslation();
 
-
 	return (
 		<div className="flex items-center justify-end space-x-5">
-      {showSearchBar &&
-        <>
-        	<HeaderSearchBar onSearch={onSearch} />
-        	<div className="h-12 my-auto border-l border-1 border-theme-primary-100" />
-        </>
-      }
+			{showSearchBar && (
+				<>
+					<HeaderSearchBar onSearch={onSearch} />
+					<div className="h-12 my-auto border-l border-1 border-theme-primary-100" />
+				</>
+			)}
 
-			<Button data-testid="contacts__add-contact-btn" className="whitespace-no-wrap" onClick={onAddContact}>{t("CONTACTS.CONTACTS_PAGE.ADD_CONTACT")}</Button>
+			<Button data-testid="contacts__add-contact-btn" className="whitespace-no-wrap" onClick={onAddContact}>
+				{t("CONTACTS.CONTACTS_PAGE.ADD_CONTACT")}
+			</Button>
 		</div>
 	);
 };
@@ -42,7 +43,7 @@ type ContactsProps = {
 };
 
 export const Contacts = ({ contacts, onSearch }: ContactsProps) => {
-  const [createIsOpen, setCreateIsOpen] = useState(false);
+	const [createIsOpen, setCreateIsOpen] = useState(false);
 
 	const { t } = useTranslation();
 
@@ -73,18 +74,18 @@ export const Contacts = ({ contacts, onSearch }: ContactsProps) => {
 		},
 	];
 
-  // TODO: replace after design phase
-  const networks = [
-    {
-      label: "Ark Ecosystem",
-      value: "ark",
-      icon: "Ark",
-    },
-  ];
+	// TODO: replace after design phase
+	const networks = [
+		{
+			label: "Ark Ecosystem",
+			value: "ark",
+			icon: "Ark",
+		},
+	];
 
-  const handleOnSave = () => {
-    setCreateIsOpen(false);
-  };
+	const handleOnSave = () => {
+		setCreateIsOpen(false);
+	};
 
 	return (
 		<div data-testid="contacts" className="flex flex-col min-h-screen -m-5 bg-theme-neutral-200">
@@ -96,12 +97,12 @@ export const Contacts = ({ contacts, onSearch }: ContactsProps) => {
 						title={t("CONTACTS.CONTACTS_PAGE.TITLE")}
 						subtitle={t("CONTACTS.CONTACTS_PAGE.SUBTITLE")}
 						extra={
-              <ContactsHeaderExtra
-                showSearchBar={contacts.length > 0}
-                onSearch={onSearch}
-                onAddContact={() => setCreateIsOpen(true)}
-              />
-            }
+							<ContactsHeaderExtra
+								showSearchBar={contacts.length > 0}
+								onSearch={onSearch}
+								onAddContact={() => setCreateIsOpen(true)}
+							/>
+						}
 					/>
 				</div>
 
@@ -116,20 +117,20 @@ export const Contacts = ({ contacts, onSearch }: ContactsProps) => {
 					{contacts.length > 0 && (
 						<div className="w-full">
 							<Table columns={listColumns} data={contacts}>
-                {(contact: any) => <ContactListItem contact={contact} />}
+								{(contact: any) => <ContactListItem contact={contact} />}
 							</Table>
 						</div>
 					)}
 				</div>
 			</div>
 
-      <CreateContact
-        isOpen={createIsOpen}
-        networks={networks}
-        onCancel={() => setCreateIsOpen(false)}
-        onClose={() => setCreateIsOpen(false)}
-        onSave={handleOnSave}
-      />
+			<CreateContact
+				isOpen={createIsOpen}
+				networks={networks}
+				onCancel={() => setCreateIsOpen(false)}
+				onClose={() => setCreateIsOpen(false)}
+				onSave={handleOnSave}
+			/>
 		</div>
 	);
 };
